# app/app.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "app.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/service.app.v1.App/GetAppSecret": {
            "get": {
                "summary": "/service.app.v1.App/GetAppSecret",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.service.app.v1.AppReply"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "sappkey",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "service.app.v1.App"
                ]
            }
        },
        "/service.app.v1.App/GetAppSkipUrls": {
            "get": {
                "summary": "/service.app.v1.App/GetAppSkipUrls",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.service.app.v1.SkipUrlsReply"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "sappkey",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "service.app.v1.App"
                ]
            }
        },
        "/service.app.v1.App/Ping": {
            "get": {
                "summary": "/service.app.v1.App/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "service.app.v1.App"
                ]
            }
        }
    },
    "definitions": {
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        },
        ".service.app.v1.AppReply": {
            "type": "object",
            "properties": {
                "app_key": {
                    "type": "string"
                },
                "app_secret": {
                    "type": "string"
                }
            }
        },
        ".service.app.v1.AppReq": {
            "type": "object",
            "properties": {
                "sappkey": {
                    "type": "string"
                }
            },
            "required": [
                "sappkey"
            ]
        },
        ".service.app.v1.SkipUrlsReply": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}